Implement QPSK TX - RX with USRP hardware
And fix some small issues in QPSK simulation.

The signal flow graph was tested with a coaxial cable, as well as with
an 2.4 GHz antenna. The constellation diagram looked OK, but the sent
data (bytes of lena512color.tiff) were neither decoded nor tested. That
is because we need to synchronize data frames with a header thing or
something like that (which is still not implemented yet).